+++ 20_Spring_WebFlux/src/test/java/ru/otus/spring/webflux/repository/BookRepositoryTest.java
@@ -1,4 +1,4 @@
-package ru.otus.pk.spring.repository;
+package ru.otus.spring.webflux.repository;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
