+++ 20_Spring_WebFlux/src/test/java/ru/otus/spring/webflux/repository/CommentRepositoryTest.java
@@ -1,4 +1,4 @@
-package ru.otus.pk.spring.repository;
+package ru.otus.spring.webflux.repository;
 
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.DisplayName;
